refactor treelike functionality out of Element into Node class
