+++ browsergui/elements/__init__.py
@@ -1,8 +1,9 @@
 import collections
 import json
-import weakref
 import xml.dom.minidom
 import xml.parsers.expat
+
+from ._node import Node
 
 _unique_id_counter = 0
 def unique_id():
@@ -28,18 +29,18 @@ def new_tag(tag_name):
 def styling_to_css(styling):
   return ' '.join('{}: {};'.format(key, value) for key, value in styling.items())
 
-class Element(object):
+class Element(Node):
   """A conceptual GUI element, like a button or a table.
 
   Elements are arranged in trees: an Element may have children (other Elements) or not, and it may have a parent or not.
   Every element has a unique identifier, accessible by the :func:`id` method.
   """
   def __init__(self, tag_name, children=()):
+    super(Element, self).__init__()
     self.tag = new_tag(tag_name)
     self.tag.setAttribute('id', unique_id())
     self._styling = {}
 
-    self.parent_weakref = None
     self.children = []
     self.callbacks = collections.defaultdict(list)
 
@@ -62,13 +63,6 @@ class Element(object):
   @property
   def id(self):
     return self.tag.getAttribute('id')
-
-  def walk(self):
-    """Iterates over the Element and all the Elements below it in the tree."""
-    yield self
-    for child in self.children:
-      for descendant in child.walk():
-        yield descendant
 
   def append(self, child):
     """Add a new child after all existing children.
@@ -113,21 +107,6 @@ class Element(object):
     self.parent.register_child(sibling)
 
   @property
-  def parent(self):
-    """
-    :returns: the element's parent, or None if the Element is orphaned
-    """
-    return (None if self.parent_weakref is None else self.parent_weakref())
-  @parent.setter
-  def parent(self, parent):
-    if parent is None:
-      self.parent_weakref = None
-    elif self.orphaned:
-      self.parent_weakref = weakref.ref(parent)
-    else:
-      raise NotOrphanedError('only orphaned elements can be given new parents')
-
-  @property
   def next_sibling(self):
     """    
     :returns: the next child of the element's parent, or None if there isn't one.
@@ -155,14 +134,6 @@ class Element(object):
   def html(self):
     """An HTML representation of the element and all its children."""
     return self.tag.toprettyxml()
-
-  @property
-  def orphaned(self):
-    """Whether the Element has no parent.
-
-    :rtype: bool
-    """
-    return (self.parent is None)
 
   @property
   def gui(self):
